+++ feature_engine/outliers/trimmer.py
@@ -4,7 +4,7 @@
 import numpy as np
 import pandas as pd
 
-from .winsorizer import WinsorizerBase
+from feature_engine.outliers.base_outlier import WinsorizerBase
 
 
 class OutlierTrimmer(WinsorizerBase):
